Support non-LE replication of the autoscaler.
This change continues the march of my prior changes to separate the awareness of how ownership of a bucket is established from how the bucket ownership information is consumed.

With this change, the `leaseTracker` is moved into its own file and no longer invoked by `statsforwarder.New`, but depending on the method of HA employed by the entrypoint the resulting `Forwarder` may be passed to either `LeaseBasedProcessor` or `StatefulSetBasedProcessor` to configure the Forwarder's processors on an ongoing basis.

This change also implements the `StatefulSetBaseProcessor`, and configures the autoscalers entrypoint to dynamically configure the appropriate HA mode based on the provided environment variables.
