+++ go.mod
@@ -36,6 +36,6 @@ require (
 	k8s.io/kube-openapi v0.0.0-20200410145947-bcb3869e6f29
 	knative.dev/caching v0.0.0-20201209181541-0e55b559628c
 	knative.dev/hack v0.0.0-20201201234937-fddbf732e450
-	knative.dev/networking v0.0.0-20201210074041-cdd0e325c1a9
-	knative.dev/pkg v0.0.0-20201210014142-0c53297607c6
+	knative.dev/networking v0.0.0-20201209181441-896c0e7c4864
+	knative.dev/pkg v0.0.0-20201209172941-9b9407a76df1
 )
